+++ src/port_diff/squash.rs
@@ -1,8 +1,5 @@
 use std::collections::{BTreeMap, BTreeSet};
 
-use bimap::BiBTreeMap;
-use derive_more::From;
-use derive_where::derive_where;
 use petgraph::visit::{EdgeRef, IntoEdges};
 use union_find::{QuickUnionUf, UnionBySize, UnionFind};